Rename XCUITestExtensions -> VSMobileCenterExtensions
