+++ VSMobileCenterExtensions.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:/Users/chrisf/test-cloud-xcuitest-extensions/XCUITestExtensions.xcodeproj">
+      location = "self:/Users/chrisf/test-cloud-xcuitest-extensions/VSMobileCenterExtensions.xcodeproj">
    </FileRef>
 </Workspace>
