+++ VSMobileCenterExtensions/Info.plist
