+++ VSMobileCenterExtensions/MCLabel.m
