+++ VSMobileCenterExtensions/VSMobileCenterExtensions.h
@@ -2,11 +2,11 @@
 #import <UIKit/UIKit.h>
 
 //! Project version number for XCUITestExtensions.
-FOUNDATION_EXPORT double XCUITestExtensionsVersionNumber;
+FOUNDATION_EXPORT double VSMobileCenterExtensionsVersionNumber;
 
 //! Project version string for XCUITestExtensions.
-FOUNDATION_EXPORT const unsigned char XCUITestExtensionsVersionString[];
+FOUNDATION_EXPORT const unsigned char VSMobileCenterExtensionsVersionString[];
 
 // In this header, you should import all the public headers of your framework using statements like
 
-#import <XCUITestExtensions/MCLabel.h>
+#import <VSMobileCenterExtensions/MCLabel.h>
