+++ VSMobileCenterExtensionsTests/Info.plist
